+++ ChangeLog.md
@@ -1,9 +1,14 @@
-* 0.2 (2018-08-17)
+# 0.3 (2019-10-09)
+- check package provides instead of rpms
+- use ByteString for packages (provides)
+- move to simple-cmd-args (optparse-applicative)
+
+# 0.2 (2018-08-17)
 - error if no spec file
 - filter out hidden files from deps search
 - add --version
 - generate manpage with help2man
 
-* 0.1 (2018-07-18)
+# 0.1 (2018-07-18)
 - initial release with sort, deps and rdeps (reverse dependencies) commands